pyproject.toml
[tool.poetry]
name = "everythingelse"
version = "0.1.0"
description = ""
authors = ["Robert Meissner <robert.meissner@codecentric.de>"]

[tool.poetry.dependencies]
python = "^3.10"
pyperclip = "^1.8.2"
odfpy = "^1.4.1"
pandas = "^1.5.2"
pandas-ods-reader = "^0.1.4"
python-dotenv = "^0.21.0"
httpx = "^0.23.1"
matplotlib = "^3.6.2"
pre-commit = "^2.20.0"
todoist-api-python = "^2.0.2"
md2notion = "^2.4.1"
google-auth = "^2.23.4"
google-auth-oauthlib = "^1.1.0"
google-auth-httplib2 = "^0.1.1"
google-api-python-client = "^2.108.0"
streamlit = "^1.37.1"
tqdm = "^4.66.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
